+++ src/components/DappBrowser/search/Search.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { SearchBar } from './SearchBar';
-import { SearchResults } from './SearchResults';
+import { SearchResults } from './search-results/SearchResults';
 
 export const Search = () => {
   return (
