+++ src/components/DappBrowser/search/search-results/SearchResults.tsx
@@ -1,42 +1,15 @@
 import { Box, Inline, Inset, Stack, Text, TextIcon } from '@/design-system';
 import React from 'react';
 import Animated, { useAnimatedStyle } from 'react-native-reanimated';
-import { ButtonPressAnimation } from '../../animations';
-import { ImgixImage } from '../../images';
-import { useBrowserContext } from '../BrowserContext';
-
-const SearchResult = ({ suggested }: { suggested?: boolean }) => {
-  return (
-    <Box as={ButtonPressAnimation} padding="8px" borderRadius={18} background={suggested ? 'fill' : undefined}>
-      <Inline space="12px" alignVertical="center">
-        <Box
-          as={ImgixImage}
-          source={{ uri: 'https://pbs.twimg.com/profile_images/1741494128779886592/RY4V0T2F_400x400.jpg' }}
-          size={48}
-          background="surfacePrimary"
-          shadow="24px"
-          width={{ custom: 40 }}
-          height={{ custom: 40 }}
-          borderRadius={10}
-        />
-        <Stack space="10px">
-          <Text size="17pt" weight="bold" color="label">
-            Uniswap
-          </Text>
-          <Text size="13pt" weight="bold" color="labelTertiary">
-            app.uniswap.org
-          </Text>
-        </Stack>
-      </Inline>
-    </Box>
-  );
-};
+import { ButtonPressAnimation } from '../../../animations';
+import { useBrowserContext } from '../../BrowserContext';
+import { SearchResult } from './SearchResult';
 
 export const SearchResults = () => {
-  const { searchViewProgress, searchInputRef, tabViewProgress } = useBrowserContext();
+  const { searchViewProgress, searchInputRef } = useBrowserContext();
   const backgroundStyle = useAnimatedStyle(() => ({
     opacity: searchViewProgress?.value,
-    pointerEvents: (tabViewProgress?.value ?? 0) < 1 ? 'box-none' : 'none',
+    pointerEvents: searchViewProgress?.value ? 'box-none' : 'none',
   }));
 
   return (
